add js code gen tests and remove dependecny checking test
the dependency checking was not super helpful since if a test failed, it was not obvious what went wrong. it probbaly is useful as some sort of unit test though, so worth revisiting at some point - or replacing with some other test on the synthesis part. The new js code gen test only tests the last syntactic transformation step. We should also add a python check and other checks for whichever targets we support